update kudu code

+++ java/kudu-mapreduce/src/test/java/org/apache/kudu/mapreduce/TestJarFinder.java
@@ -14,6 +14,7 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+
 package org.apache.kudu.mapreduce;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -35,14 +36,20 @@ import org.apache.commons.logging.LogFactory;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+
+import org.apache.kudu.test.junit.RetryRule;
 
 /**
  * This file was forked from hbase/branches/master@4ce6f48.
  */
 public class TestJarFinder {
 
-  private static File testDir;
+  private File testDir;
+
+  @Rule
+  public RetryRule retryRule = new RetryRule();
 
   @Before
   public void setUp() throws Exception {
@@ -52,7 +59,9 @@ public class TestJarFinder {
 
   @After
   public void tearDown() throws Exception {
-    FileUtils.deleteDirectory(testDir);
+    if (testDir != null) {
+      FileUtils.deleteDirectory(testDir);
+    }
   }
 
   @Test
@@ -72,44 +81,40 @@ public class TestJarFinder {
 
   @Test
   public void testExistingManifest() throws Exception {
-    File dir = new File(testDir,
-      TestJarFinder.class.getName() + "-testExistingManifest");
-    File metaInfDir = new File(dir, "META-INF");
-    metaInfDir.mkdirs();
-    File manifestFile = new File(metaInfDir, "MANIFEST.MF");
-    Manifest manifest = new Manifest();
-    OutputStream os = new FileOutputStream(manifestFile);
-    manifest.write(os);
-    os.close();
-
-    File propsFile = new File(dir, "props.properties");
-    Writer writer = Files.newBufferedWriter(propsFile.toPath(), UTF_8);
-    new Properties().store(writer, "");
-    writer.close();
-    ByteArrayOutputStream baos = new ByteArrayOutputStream();
-    JarOutputStream zos = new JarOutputStream(baos);
-    JarFinder.jarDir(dir, "", zos);
-    JarInputStream jis =
-      new JarInputStream(new ByteArrayInputStream(baos.toByteArray()));
-    Assert.assertNotNull(jis.getManifest());
-    jis.close();
+    File dir = new File(testDir, TestJarFinder.class.getName() + "-testExistingManifest");
+    Assert.assertTrue(dir.mkdirs());
+    writeManifest(dir);
+    Assert.assertNotNull(getManifest(dir));
   }
 
   @Test
   public void testNoManifest() throws Exception {
-    File dir = new File(testDir,
-      TestJarFinder.class.getName() + "-testNoManifest");
-    dir.mkdirs();
+    File dir = new File(testDir, TestJarFinder.class.getName() + "-testNoManifest");
+    Assert.assertTrue(dir.mkdirs());
+    Assert.assertNotNull(getManifest(dir));
+  }
+
+  private void writeManifest(File dir) throws Exception {
+    File metaInfDir = new File(dir, "META-INF");
+    Assert.assertTrue(metaInfDir.mkdirs());
+    File manifestFile = new File(metaInfDir, "MANIFEST.MF");
+    Manifest manifest = new Manifest();
+    try (OutputStream os = new FileOutputStream(manifestFile)) {
+      manifest.write(os);
+    }
+  }
+
+  private Manifest getManifest(File dir) throws Exception {
     File propsFile = new File(dir, "props.properties");
-    Writer writer = Files.newBufferedWriter(propsFile.toPath(), UTF_8);
-    new Properties().store(writer, "");
-    writer.close();
-    ByteArrayOutputStream baos = new ByteArrayOutputStream();
-    JarOutputStream zos = new JarOutputStream(baos);
-    JarFinder.jarDir(dir, "", zos);
-    JarInputStream jis =
-      new JarInputStream(new ByteArrayInputStream(baos.toByteArray()));
-    Assert.assertNotNull(jis.getManifest());
-    jis.close();
+    try (Writer writer = Files.newBufferedWriter(propsFile.toPath(), UTF_8)) {
+      new Properties().store(writer, "");
+    }
+    try (ByteArrayOutputStream baos = new ByteArrayOutputStream();
+         JarOutputStream zos = new JarOutputStream(baos)) {
+      JarFinder.jarDir(dir, "", zos);
+      try (JarInputStream jis = new JarInputStream(new ByteArrayInputStream(baos.toByteArray()))) {
+        return jis.getManifest();
+      }
+    }
   }
 }
